+++ src/OpenSpaceToolkit/Physics/Environment/Object/Celestial/Sun.cpp
@@ -5,7 +5,7 @@
 
 #include <OpenSpaceToolkit/Physics/Coordinate/Frame.hpp>
 #include <OpenSpaceToolkit/Physics/Coordinate/Frame/Provider/Static.hpp>
-#include <OpenSpaceToolkit/Physics/Environment/Ephemeris/CelestialAnalytical.hpp>
+#include <OpenSpaceToolkit/Physics/Environment/Ephemeris/Analytical.hpp>
 #include <OpenSpaceToolkit/Physics/Environment/Ephemeris/SPICE.hpp>
 #include <OpenSpaceToolkit/Physics/Environment/Object/Celestial/Sun.hpp>
 
@@ -62,10 +62,10 @@ Sun Sun::Spherical()
 
 Sun Sun::Analytical()
 {
-    using ostk::physics::environment::ephemeris::CelestialAnalytical;
+    using ostk::physics::environment::ephemeris::Analytical;
 
     return {
-        std::make_shared<CelestialAnalytical>(Celestial::Type::Sun),
+        std::make_shared<Analytical>(Frame::SunAnalytical()),
         std::make_shared<SunGravitationalModel>(SunGravitationalModel::Type::Spherical),
     };
 }
