+++ test/OpenSpaceToolkit/Physics/Coordinate/Frame/Provider/MoonAnalytical.test.cpp
@@ -1,0 +1,169 @@
+/// Apache License 2.0
+
+#include <cmath>
+
+#include <OpenSpaceToolkit/Core/Container/Array.hpp>
+#include <OpenSpaceToolkit/Core/Type/Real.hpp>
+#include <OpenSpaceToolkit/Core/Type/Shared.hpp>
+#include <OpenSpaceToolkit/Core/Type/String.hpp>
+
+#include <OpenSpaceToolkit/Physics/Coordinate/Frame.hpp>
+#include <OpenSpaceToolkit/Physics/Coordinate/Frame/Provider/MoonAnalytical.hpp>
+#include <OpenSpaceToolkit/Physics/Environment/Ephemeris/SPICE/Engine.hpp>
+#include <OpenSpaceToolkit/Physics/Environment/Ephemeris/SPICE/Manager.hpp>
+#include <OpenSpaceToolkit/Physics/Environment/Object/Celestial/Moon.hpp>
+#include <OpenSpaceToolkit/Physics/Time/DateTime.hpp>
+#include <OpenSpaceToolkit/Physics/Time/Instant.hpp>
+#include <OpenSpaceToolkit/Physics/Time/Scale.hpp>
+
+#include <Global.test.hpp>
+
+using ostk::core::container::Array;
+using ostk::core::filesystem::Directory;
+using ostk::core::filesystem::File;
+using ostk::core::filesystem::Path;
+using ostk::core::type::Real;
+using ostk::core::type::Shared;
+using ostk::core::type::String;
+
+using ostk::mathematics::object::Vector3d;
+
+using ostk::physics::coordinate::Frame;
+using ostk::physics::coordinate::frame::provider::MoonAnalytical;
+using ostk::physics::environment::ephemeris::spice::Engine;
+using ostk::physics::environment::ephemeris::spice::Kernel;
+using ostk::physics::environment::ephemeris::spice::Manager;
+using ostk::physics::environment::object::celestial::Moon;
+using ostk::physics::time::DateTime;
+using ostk::physics::time::Instant;
+using ostk::physics::time::Scale;
+
+class OpenSpaceToolkit_Physics_Coordinate_Frame_Provider_MoonAnalytical : public ::testing::Test
+{
+   protected:
+    Engine& engine_;
+    Manager& manager_;
+
+    OpenSpaceToolkit_Physics_Coordinate_Frame_Provider_MoonAnalytical()
+        : engine_(Engine::Get()),
+          manager_(Manager::Get())
+    {
+    }
+
+    void TearDown() override
+    {
+        engine_.reset();
+        manager_.setMode(Manager::Mode::Automatic);
+    }
+
+    void loadSpiceKernels()
+    {
+        const Directory spiceLocalRepository =
+            Directory::Path(Path::Parse("/app/test/OpenSpaceToolkit/Physics/Environment/Ephemeris/SPICE"));
+
+        manager_.setMode(Manager::Mode::Manual);
+
+        engine_.reset();
+
+        engine_.loadKernel(Kernel::File(File::Path(spiceLocalRepository.getPath() + Path::Parse("naif0012.tls"))));
+        engine_.loadKernel(Kernel::File(File::Path(spiceLocalRepository.getPath() + Path::Parse("de430.bsp"))));
+        engine_.loadKernel(Kernel::File(File::Path(spiceLocalRepository.getPath() + Path::Parse("pck00010.tpc"))));
+    }
+
+    static Array<Instant> ComparisonInstants()
+    {
+        return {
+            Instant::DateTime(DateTime(2020, 1, 1, 0, 0, 0), Scale::UTC),
+            Instant::DateTime(DateTime(2020, 8, 15, 6, 30, 0), Scale::UTC),
+            Instant::DateTime(DateTime(2021, 3, 10, 12, 0, 0), Scale::UTC),
+            Instant::DateTime(DateTime(2021, 11, 25, 18, 45, 0), Scale::UTC),
+            Instant::DateTime(DateTime(2022, 6, 1, 3, 15, 0), Scale::UTC),
+            Instant::DateTime(DateTime(2023, 1, 20, 9, 0, 0), Scale::UTC),
+            Instant::DateTime(DateTime(2023, 9, 5, 21, 30, 0), Scale::UTC),
+            Instant::DateTime(DateTime(2024, 4, 14, 15, 0, 0), Scale::UTC),
+            Instant::DateTime(DateTime(2025, 2, 28, 0, 0, 0), Scale::UTC),
+            Instant::DateTime(DateTime(2025, 12, 15, 12, 0, 0), Scale::UTC),
+            Instant::DateTime(DateTime(2026, 5, 30, 6, 0, 0), Scale::UTC),
+        };
+    }
+
+    static Real ComputeAngularSeparation_deg(const Vector3d& aFirstVector, const Vector3d& aSecondVector)
+    {
+        const double cosine = std::clamp(aFirstVector.normalized().dot(aSecondVector.normalized()), -1.0, 1.0);
+
+        return std::acos(cosine) * 180.0 / M_PI;
+    }
+};
+
+TEST_F(OpenSpaceToolkit_Physics_Coordinate_Frame_Provider_MoonAnalytical, IsDefined)
+{
+    {
+        EXPECT_TRUE(MoonAnalytical().isDefined());
+    }
+}
+
+TEST_F(OpenSpaceToolkit_Physics_Coordinate_Frame_Provider_MoonAnalytical, GetTransformAt)
+{
+    {
+        const Instant instant = Instant::DateTime(DateTime(2024, 1, 1, 0, 0, 0), Scale::UTC);
+
+        EXPECT_TRUE(MoonAnalytical().getTransformAt(instant).isDefined());
+    }
+
+    {
+        EXPECT_ANY_THROW(MoonAnalytical().getTransformAt(Instant::Undefined()));
+    }
+}
+
+TEST_F(OpenSpaceToolkit_Physics_Coordinate_Frame_Provider_MoonAnalytical, PositionVersusSpice)
+{
+    {
+        this->loadSpiceKernels();
+
+        const Shared<const Frame> analyticalFrameSPtr = Frame::MoonAnalytical();
+
+        const Moon moonSpice = Moon::Default();
+
+        for (const auto& instant : ComparisonInstants())
+        {
+            const Vector3d x_GCRF_analytical =
+                analyticalFrameSPtr->getOriginIn(Frame::GCRF(), instant).getCoordinates();
+            const Vector3d x_GCRF_spice = moonSpice.getPositionIn(Frame::GCRF(), instant).getCoordinates();
+
+            const Real angularSeparation_deg = ComputeAngularSeparation_deg(x_GCRF_analytical, x_GCRF_spice);
+            const Real relativeDistanceError =
+                std::abs(x_GCRF_analytical.norm() - x_GCRF_spice.norm()) / x_GCRF_spice.norm();
+
+            EXPECT_LT(angularSeparation_deg, 0.5) << String::Format(
+                "{}: angular separation = {} [deg]", instant.toString(), angularSeparation_deg.toString()
+            );
+            EXPECT_LT(relativeDistanceError, 0.015) << String::Format(
+                "{}: relative distance error = {}", instant.toString(), relativeDistanceError.toString()
+            );
+        }
+    }
+}
+
+TEST_F(OpenSpaceToolkit_Physics_Coordinate_Frame_Provider_MoonAnalytical, VelocityVersusSpice)
+{
+    {
+        this->loadSpiceKernels();
+
+        const Shared<const Frame> analyticalFrameSPtr = Frame::MoonAnalytical();
+
+        const Moon moonSpice = Moon::Default();
+
+        for (const auto& instant : ComparisonInstants())
+        {
+            const Vector3d v_GCRF_analytical =
+                analyticalFrameSPtr->getVelocityIn(Frame::GCRF(), instant).getCoordinates();
+            const Vector3d v_GCRF_spice = moonSpice.getVelocityIn(Frame::GCRF(), instant).getCoordinates();
+
+            const Real relativeVelocityError = (v_GCRF_analytical - v_GCRF_spice).norm() / v_GCRF_spice.norm();
+
+            EXPECT_LT(relativeVelocityError, 0.05) << String::Format(
+                "{}: relative velocity error = {}", instant.toString(), relativeVelocityError.toString()
+            );
+        }
+    }
+}
