+++ test/OpenSpaceToolkit/Physics/Environment/Object/Celestial/Moon.test.cpp
@@ -19,6 +19,7 @@
 #include <OpenSpaceToolkit/Mathematics/Geometry/3D/Transformation/Rotation/Quaternion.hpp>
 #include <OpenSpaceToolkit/Mathematics/Geometry/3D/Transformation/Rotation/RotationVector.hpp>
 
+#include <OpenSpaceToolkit/Physics/Coordinate/Frame.hpp>
 #include <OpenSpaceToolkit/Physics/Coordinate/Spherical/LLA.hpp>
 #include <OpenSpaceToolkit/Physics/Environment.hpp>
 #include <OpenSpaceToolkit/Physics/Environment/Object/Celestial/Moon.hpp>
@@ -31,7 +32,11 @@
 
 #include <Global.test.hpp>
 
+using ostk::physics::coordinate::Frame;
 using ostk::physics::environment::object::celestial::Moon;
+using ostk::physics::time::DateTime;
+using ostk::physics::time::Instant;
+using ostk::physics::time::Scale;
 
 // TEST (OpenSpaceToolkit_Physics_Environment_Object_Celestial_Moon, Constructor)
 // {
@@ -50,5 +55,20 @@ TEST(OpenSpaceToolkit_Physics_Environment_Object_Celestial_Moon, Spherical)
 {
     {
         EXPECT_NO_THROW(Moon::Spherical());
+    }
+}
+
+TEST(OpenSpaceToolkit_Physics_Environment_Object_Celestial_Moon, Analytical)
+{
+    {
+        EXPECT_NO_THROW(Moon::Analytical());
+    }
+
+    {
+        const Moon moon = Moon::Analytical();
+
+        const Instant instant = Instant::DateTime(DateTime(2024, 1, 1, 0, 0, 0), Scale::UTC);
+
+        EXPECT_TRUE(moon.getPositionIn(Frame::GCRF(), instant).isDefined());
     }
 }
